// src/app/models/product.ts
export interface ProductApiResponse {
  message: string;
  totalItems: number;
  data: Product[];
}

/**
 * 商品
 */
export interface Product {
  id: number;
  name: string;
  name_kana: string;
  supplier_id: number;
  supplier_name: string;
  supplier_product_id: number;
  supplier_product_cd: string;
  supplier_sales_tax_division_id: number;
  supplier_sales_fraction_division_id: number;
  supplier_cost_price: number;
  b_supplier_cost_price: number;
  c_supplier_cost_price: number;
  standard: string;
  part_number: string;
  sales_tax_division_id: number;
  sales_fraction_division_id: number;
  barcode: string;
  b_barcode: string;
  c_barcode: string;
  quantity: number | '';
  b_quantity: number | '';
  c_quantity: number | '';
  gross_profit_rate: number | '';
  b_gross_profit_rate: number | '';
  c_gross_profit_rate: number | '';
  selling_price: number;
  b_selling_price: number | '';
  c_selling_price: number | '';
  discount_division_id: number | '';
  unit_division_id: number | '';
  b_unit_division_id: number | '';
  c_unit_division_id: number | '';
  product_division_id: number | '';
  point_division_id: number;
  issuance_warranty_division_id: number;
  large_category_id: number | '';
  medium_category_id: number | '';
  small_category_id: number | '';
  remarks_1: string;
  remarks_2: string;
  remarks_3: string;
  remarks_4: string;
  seal_print_division_id: number;
  price_tag_division_id: number;
  ordering_target_barcode: string;
  regulated_stock_num: number | '';
  ordering_point: number | '';
  order_lead_time: number | '';
  shelf_division_id: number;
  shelf_col_division_id: number;
  image_path: string;
  custom_large_category_id: number | '';
  custom_medium_category_id: number | '';
  custom_small_category_id: number | '';
  custom_tag: string;
  custom_tag_id: string | '';
  data_permission_division_id: number | '';
  basic_shipping_fee: number;
  shipping_surcharge: number;
  minimum_order_quantity: number;
  order_unit: string;
  order_cost_price: number;
  seal_price_indication: number | '';
  created_at: string;
  created_id: number;
  updated_at: string;
  updated_id: number;
  out_of_service_at: string;
  out_of_service_id: number | '';
  division_supplier_sales_tax_name: string;
  division_supplier_sales_tax_code: number;
  division_supplier_sales_tax_value: string;
  division_supplier_sales_fraction_name: string;
  division_supplier_sales_fraction_code: number;
  division_supplier_sales_fraction_value: string;
  division_supplier_product_status_value: string;
  division_sales_tax_name: string;
  division_sales_tax_code: number;
  division_sales_tax_value: string;
  division_sales_fraction_name: string;
  division_sales_fraction_code: number;
  division_sales_fraction_value: string;
  division_discount_name: string;
  division_discount_code: number;
  division_discount_value: string;
  division_point_name: string;
  division_point_code: number;
  division_point_value: string;
  division_issuance_warranty_name: string;
  division_issuance_warranty_code: number;
  division_issuance_warranty_value: string;
  division_seal_print_name: string;
  division_seal_print_code: number;
  division_seal_print_value: string;
  division_price_tag_name: string;
  division_price_tag_code: number;
  division_price_tag_value: string;
  division_shelf_name: string;
  division_shelf_code: number;
  division_shelf_value: string;
  division_shelf_col_name: string;
  division_shelf_col_code: number;
  division_shelf_col_value: string;
  division_unit_name: string;
  division_unit_code: number | '';
  division_unit_value: string;
  division_b_unit_name: string;
  division_b_unit_code: number | '';
  division_b_unit_value: string;
  division_c_unit_name: string;
  division_c_unit_code: number | '';
  division_c_unit_value: string;
  division_product_name: string;
  division_product_code: number;
  division_product_value: string;
  large_category_name: string;
  medium_category_name: string;
  small_category_name: string;
  custom_large_category_name: string;
  custom_medium_category_name: string;
  custom_small_category_name: string;
  division_data_permission_name: string;
  division_data_permission_code: number;
  division_data_permission_value: string;
  employee_created_last_name: string;
  employee_created_first_name: string;
  employee_updated_last_name: string;
  employee_updated_first_name: string;
  employee_out_of_service_last_name: string;
  employee_out_of_service_first_name: string;
}